Implement working endpoint

+++ src/main/java/com/backend/challenge/PlaceResult.java
@@ -1,4 +1,4 @@
 package com.backend.challenge;
 
-public record PlaceResult(String name, String address, long distance) {
+public record PlaceResult(String name, String address, double distance) {
 }
